requirements.txt
# requirements.txt - Pure Gemini Implementation
# Core dependencies
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
google-generativeai==0.3.1

# LiveKit and voice processing
livekit-api==1.2.0
livekit-rtc==0.7.0
edge-tts==6.1.9

# LiveKit plugins
deepgram-sdk==2.14.0

# Testing and utilities
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0
python-multipart==0.0.6